requirements.txt
openpyxl
pygame
memoization
numpy
parameterized
argh
PyYAML
jsonpickle